chore: move tests to remove circular dep
