+++ packages/tests/.eslintrc.cjs
@@ -3,4 +3,7 @@ module.exports = {
     tsconfigRootDir: __dirname,
     project: "./tsconfig.json",
   },
+  rules: {
+    "@typescript-eslint/no-non-null-assertion": "off",
+  },
 };
