+++ packages/tests/package.json
@@ -43,12 +43,20 @@
   "scripts": {
     "fix": "run-s fix:*",
     "fix:prettier": "prettier . --write",
-    "fix:lint": "eslint tests --ext .ts --ext .cjs --fix",
+    "fix:lint": "eslint src tests --ext .ts --ext .cjs --fix",
+    "pretest": "run-s pretest:*",
+    "pretest:1-init-git-submodules": "[ -f '../../nwaku/build/wakunode2' ] || git submodule update --init --recursive",
+    "pretest:2-build-nwaku": "[ -f '../../nwaku/build/wakunode2' ] || run-s nwaku:build",
+    "nwaku:build": "(PROC=$(nproc --all 2>/dev/null || echo 2); cd ../../nwaku; make -j$PROC update; NIMFLAGS=\"-d:chronicles_colors=off -d:chronicles_sinks=textlines -d:chronicles_log_level=TRACE\" make -j$PROC wakunode2)",
+    "nwaku:force-build": "(cd ../../nwaku && rm -rf ./build/ ./vendor) && run-s nwaku:build",
     "check": "run-s check:*",
-    "check:lint": "eslint tests --ext .ts",
+    "check:lint": "eslint src tests --ext .ts",
     "check:prettier": "prettier . --list-different",
-    "check:spelling": "cspell \"{README.md,src/**/*.ts}\"",
+    "check:spelling": "cspell \"{README.md,{tests,src}/**/*.ts}\"",
     "check:tsc": "tsc -p tsconfig.json",
+    "test": "run-s test:*",
+    "test:node": "TS_NODE_PROJECT=./tsconfig.json mocha",
+    "test:browser": "karma start karma.conf.cjs",
     "reset-hard": "git clean -dfx -e .idea && git reset --hard && npm i && npm run build"
   },
   "engines": {
