+++ packages/tests/tests/enr.node.spec.ts
@@ -1,12 +1,11 @@
+import { ENR } from "@waku/core/lib/enr";
+import { waitForRemotePeer } from "@waku/core/lib/wait_for_remote_peer";
 import { createPrivacyNode } from "@waku/create";
 import type { WakuPrivacy } from "@waku/interfaces";
 import { Protocols } from "@waku/interfaces";
 import { expect } from "chai";
 
-import { makeLogFileName, NOISE_KEY_1, Nwaku } from "../../test_utils";
-import { waitForRemotePeer } from "../wait_for_remote_peer";
-
-import { ENR } from "./enr";
+import { makeLogFileName, NOISE_KEY_1, Nwaku } from "../src";
 
 describe("ENR Interop: nwaku", function () {
   let waku: WakuPrivacy;
@@ -32,6 +31,8 @@ describe("ENR Interop: nwaku", function () {
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore TODO: uniformize usage of multiaddr lib across repos
     await waku.dial(multiAddrWithId);
     await waitForRemotePeer(waku, [Protocols.Relay]);
 
@@ -64,6 +65,8 @@ describe("ENR Interop: nwaku", function () {
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore TODO: uniformize usage of multiaddr lib across repos
     await waku.dial(multiAddrWithId);
     await waitForRemotePeer(waku, [Protocols.Relay]);
 
@@ -96,6 +99,8 @@ describe("ENR Interop: nwaku", function () {
       staticNoiseKey: NOISE_KEY_1,
     });
     await waku.start();
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore TODO: uniformize usage of multiaddr lib across repos
     await waku.dial(multiAddrWithId);
     await waitForRemotePeer(waku, [Protocols.Relay]);
